modify s_vector.f90, try to implement s_sum_x() subroutines, not finished

+++ src/common/s_vector.f90
@@ -154,7 +154,7 @@
 
 ! external arguments
 ! size of array v
-     real(dp), intent(in)  :: n
+     integer, intent(in)   :: n
 
 ! sum of array v
      real(dp), intent(out) :: vsum
